Move shared eslint base config to repo root
Per review on PR #1449: philosophically the base config that every package's
eslint config extends belongs at the monorepo root, not inside one specific
package (packages/shared). Moves packages/shared/eslint.base.mjs to
eslint.base.mjs and updates the three per-package configs' relative imports.

(ESLint flat config has no cross-package "look at the root first" resolution
to hook into here — lint-staged/lint:diff invoke each package with an
explicit --config <path>, so this is a straight file relocation + import
update, not config auto-discovery.)
